# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LeadInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 8 */
  phone: string;
  email: string;
}

export type LeadCreatedEmailDelivery = typeof LeadCreatedEmailDelivery[keyof typeof LeadCreatedEmailDelivery];


export const LeadCreatedEmailDelivery = {
  sent: 'sent',
  failed: 'failed',
} as const;

export interface LeadCreated {
  id: string;
  email: string;
  emailDelivery: LeadCreatedEmailDelivery;
  createdAt: string;
}

export type EmailTriggeredEmailDelivery = typeof EmailTriggeredEmailDelivery[keyof typeof EmailTriggeredEmailDelivery];


export const EmailTriggeredEmailDelivery = {
  sent: 'sent',
  failed: 'failed',
} as const;

export interface EmailTriggered {
  success: boolean;
  emailDelivery: EmailTriggeredEmailDelivery;
}

export interface ConsultationDay {
  date: string;
  label: string;
  times: string[];
}

export interface ConsultationInput {
  leadId: string;
  scheduledDate: string;
  /** @minLength 1 */
  scheduledTime: string;
}

export type ConsultationStatus = typeof ConsultationStatus[keyof typeof ConsultationStatus];


export const ConsultationStatus = {
  booked: 'booked',
} as const;

export interface Consultation {
  id: string;
  leadId: string;
  scheduledDate: string;
  scheduledTime: string;
  status: ConsultationStatus;
  calBookingId?: string | null;
  createdAt: string;
}