# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: giveaway
    description: Salam Journey giveaway and consultation booking
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /leads:
    post:
      operationId: createLead
      tags: [giveaway]
      summary: Register a giveaway lead
      description: Saves a lead and attempts to send the ebook using transactional email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Lead saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadCreated"
        "400":
          description: Invalid lead details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /leads/{id}/send-email:
    post:
      operationId: triggerGiveawayEmail
      tags: [giveaway]
      summary: Send the giveaway email to a registered lead
      description: Fetches the lead and sends the ebook and schedules follow-up emails.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Email sending triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTriggered"
        "404":
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /consultation-slots:
    get:
      operationId: listConsultationSlots
      tags: [giveaway]
      summary: List available consultation slots
      responses:
        "200":
          description: Available days and times
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConsultationDay"
  /consultations:
    post:
      operationId: createConsultation
      tags: [giveaway]
      summary: Book a free consultation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultationInput"
      responses:
        "201":
          description: Consultation booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
        "400":
          description: Invalid booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Slot already booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    LeadInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        phone:
          type: string
          minLength: 8
        email:
          type: string
      required:
        - name
        - phone
        - email
    LeadCreated:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        emailDelivery:
          type: string
          enum: [sent, failed]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - emailDelivery
        - createdAt
    EmailTriggered:
      type: object
      properties:
        success:
          type: boolean
        emailDelivery:
          type: string
          enum: [sent, failed]
      required:
        - success
        - emailDelivery
    ConsultationDay:
      type: object
      properties:
        date:
          type: string
          format: date
        label:
          type: string
        times:
          type: array
          items:
            type: string
      required:
        - date
        - label
        - times
    ConsultationInput:
      type: object
      properties:
        leadId:
          type: string
        scheduledDate:
          type: string
          format: date
        scheduledTime:
          type: string
          minLength: 1
      required:
        - leadId
        - scheduledDate
        - scheduledTime
    Consultation:
      type: object
      properties:
        id:
          type: string
        leadId:
          type: string
        scheduledDate:
          type: string
          format: date
        scheduledTime:
          type: string
        status:
          type: string
          enum: [booked]
        calBookingId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - leadId
        - scheduledDate
        - scheduledTime
        - status
        - createdAt