# lib/api-zod/src/generated/types/consultation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ConsultationStatus } from './consultationStatus';

export interface Consultation {
  id: string;
  leadId: string;
  scheduledDate: Date;
  scheduledTime: string;
  status: ConsultationStatus;
  calBookingId?: string | null;
  createdAt: Date;
}