# lib/api-zod/src/generated/types/consultationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ConsultationInput {
  leadId: string;
  scheduledDate: Date;
  /** @minLength 1 */
  scheduledTime: string;
}