# lib/api-zod/src/generated/types/emailTriggered.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EmailTriggeredEmailDelivery } from './emailTriggeredEmailDelivery';

export interface EmailTriggered {
  success: boolean;
  emailDelivery: EmailTriggeredEmailDelivery;
}