# lib/api-zod/src/generated/types/leadCreated.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LeadCreatedEmailDelivery } from './leadCreatedEmailDelivery';

export interface LeadCreated {
  id: string;
  email: string;
  emailDelivery: LeadCreatedEmailDelivery;
  createdAt: Date;
}